--- src/main/java/com/eureka/mindbloom/auth/exception/DuplicationEmailException.java
@@ -1,9 +1,0 @@
-package com.eureka.mindbloom.auth.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class DuplicationEmailException extends BaseException{
-    public DuplicationEmailException(String message) {
-        super(message, HttpStatus.CONFLICT);
-    }
-}
