+++ src/main/java/com/eureka/mindbloom/common/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.eureka.mindbloom.auth.exception;
+package com.eureka.mindbloom.common.exception;
 
 import com.eureka.mindbloom.common.dto.ApiResponse;
 import jakarta.servlet.http.HttpServletRequest;
