+++ src/main/java/com/eureka/mindbloom/member/exception/DuplicationEmailException.java
@@ -1,0 +1,10 @@
+package com.eureka.mindbloom.member.exception;
+
+import com.eureka.mindbloom.common.exception.BaseException;
+import org.springframework.http.HttpStatus;
+
+public class DuplicationEmailException extends BaseException {
+    public DuplicationEmailException(String message) {
+        super(message, HttpStatus.CONFLICT);
+    }
+}
